Deduplicate CPI
