+++ chains/solana/contracts/programs/ccip-offramp/src/instructions/v1/mod.rs
@@ -14,3 +14,4 @@ mod messages;
 mod ocr3base;
 mod ocr3impl;
 mod pools;
+mod rmn;
