+++ chains/solana/gobindings/ccip_offramp/UpdateReferenceAddresses.go
@@ -17,7 +17,10 @@ import (
 // # Arguments
 //
 // * `ctx` - The context containing the accounts required for updating the reference addresses.
-// * `reference_addresses` - The new reference addresses to be set.
+// * `router` - The router address to be set.
+// * `fee_quoter` - The fee_quoter address to be set.
+// * `offramp_lookup_table` - The offramp_lookup_table address to be set.
+// * `rmn_remote` - The rmn_remote address to be set.
 type UpdateReferenceAddresses struct {
 	Router             *ag_solanago.PublicKey
 	FeeQuoter          *ag_solanago.PublicKey
